+++ bindings/core/src/method/wallet.rs
@@ -29,6 +29,7 @@ use crate::OmittedDebug;
 #[derive(Clone, Derivative, Serialize, Deserialize)]
 #[derivative(Debug)]
 #[serde(tag = "name", content = "data", rename_all = "camelCase")]
+#[non_exhaustive]
 pub enum WalletMethod {
     /// Creates an account.
     /// Expected response: [`Account`](crate::Response::Account)