+++ sdk/examples/client/output/foundry.rs
@@ -27,7 +27,7 @@ use iota_sdk::{
             AliasId, AliasOutputBuilder, BasicOutputBuilder, FoundryId, FoundryOutputBuilder, NativeToken, Output,
             OutputId, SimpleTokenScheme, TokenId, TokenScheme,
         },
-        payload::{transaction::TransactionEssence, Payload},
+        payload::Payload,
     },
 };
 
@@ -274,8 +274,7 @@ async fn main() -> Result<()> {
 fn get_alias_output_id(payload: &Payload) -> Result<OutputId> {
     match payload {
         Payload::Transaction(tx_payload) => {
-            let TransactionEssence::Regular(regular) = tx_payload.essence();
-            for (index, output) in regular.outputs().iter().enumerate() {
+            for (index, output) in tx_payload.essence().as_regular().outputs().iter().enumerate() {
                 if let Output::Alias(_alias_output) = output {
                     return Ok(OutputId::new(tx_payload.id(), index.try_into().unwrap())?);
                 }
@@ -290,8 +289,7 @@ fn get_alias_output_id(payload: &Payload) -> Result<OutputId> {
 fn get_foundry_output_id(payload: &Payload) -> Result<OutputId> {
     match payload {
         Payload::Transaction(tx_payload) => {
-            let TransactionEssence::Regular(regular) = tx_payload.essence();
-            for (index, output) in regular.outputs().iter().enumerate() {
+            for (index, output) in tx_payload.essence().as_regular().outputs().iter().enumerate() {
                 if let Output::Foundry(_foundry_output) = output {
                     return Ok(OutputId::new(tx_payload.id(), index.try_into().unwrap())?);
                 }
@@ -306,8 +304,7 @@ fn get_foundry_output_id(payload: &Payload) -> Result<OutputId> {
 fn get_basic_output_id_with_native_tokens(payload: &Payload) -> Result<OutputId> {
     match payload {
         Payload::Transaction(tx_payload) => {
-            let TransactionEssence::Regular(regular) = tx_payload.essence();
-            for (index, output) in regular.outputs().iter().enumerate() {
+            for (index, output) in tx_payload.essence().as_regular().outputs().iter().enumerate() {
                 if let Output::Basic(basic_output) = output {
                     if !basic_output.native_tokens().is_empty() {
                         return Ok(OutputId::new(tx_payload.id(), index.try_into().unwrap())?);