+++ sdk/examples/client/output/nft.rs
@@ -18,7 +18,7 @@ use iota_sdk::{
         output::{
             unlock_condition::AddressUnlockCondition, BasicOutputBuilder, NftId, NftOutputBuilder, Output, OutputId,
         },
-        payload::{transaction::TransactionEssence, Payload},
+        payload::Payload,
     },
 };
 
@@ -150,8 +150,7 @@ async fn main() -> Result<()> {
 fn get_nft_output_id(payload: &Payload) -> Result<OutputId> {
     match payload {
         Payload::Transaction(tx_payload) => {
-            let TransactionEssence::Regular(regular) = tx_payload.essence();
-            for (index, output) in regular.outputs().iter().enumerate() {
+            for (index, output) in tx_payload.essence().as_regular().outputs().iter().enumerate() {
                 if let Output::Nft(_nft_output) = output {
                     return Ok(OutputId::new(tx_payload.id(), index.try_into().unwrap())?);
                 }