+++ sdk/src/wallet/account/operations/output_claiming.rs
@@ -23,6 +23,7 @@ use crate::{
 
 /// Enum to specify which outputs should be claimed
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Serialize, Deserialize)]
+#[non_exhaustive]
 pub enum OutputsToClaim {
     MicroTransactions,
     NativeTokens,